+++ src/main/java/com/example/slms/Service/StudentService.java
@@ -8,7 +8,7 @@ public interface StudentService {
     void createStudent(Student student);
     List<Student> getAllStudents();
     Student getStudent(long id);
-    void updateStudent(int id, Student student);
+    void updateStudent(long id, Student student);
     void deleteStudent(long id);
     void updateStudentSubject(long stuId, long subId);
 }
